// Router-main/src/types.ts
export interface Ticket {
  id: string; // #TRK-xxxx
  title: string;
  destinatario: string;
  address: string;
  ageMinutes: number;
  priority?: string;
  status: 'Pendente' | 'Re-agendado' | 'Devolvido' | 'Troca Motorista';
  icon: string;
}